Revert "Move HistoryManager, HistoryRecord to history package"
This reverts commit fcb914e5e667e5fb7977d529b7086bd0d0ef6880.
