+++ src/main/java/cs/f10/t1/nursetraverse/logic/commands/UndoCommand.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import cs.f10.t1.nursetraverse.commons.core.index.Index;
 import cs.f10.t1.nursetraverse.logic.commands.exceptions.CommandException;
-import cs.f10.t1.nursetraverse.model.history.HistoryRecord;
+import cs.f10.t1.nursetraverse.model.HistoryRecord;
 import cs.f10.t1.nursetraverse.model.Model;
 import javafx.collections.ObservableList;
