+++ src/main/java/cs/f10/t1/nursetraverse/ui/HistoryPanel.java
@@ -1,7 +1,7 @@
 package cs.f10.t1.nursetraverse.ui;
 
 import cs.f10.t1.nursetraverse.commons.core.index.Index;
-import cs.f10.t1.nursetraverse.model.history.HistoryRecord;
+import cs.f10.t1.nursetraverse.model.HistoryRecord;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
