+++ src/test/java/cs/f10/t1/nursetraverse/model/HistoryRecordTest.java
@@ -3,7 +3,6 @@ package cs.f10.t1.nursetraverse.model;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
-import cs.f10.t1.nursetraverse.model.history.HistoryRecord;
 import org.junit.jupiter.api.Test;
 
 import cs.f10.t1.nursetraverse.commons.core.index.Index;
